StatusNotifier: better handling of Ayatana
Ayatana indicators don't provide their service when registering
an item. Instead they provide their object path. The SNI widget
deals with this by emplying a message handler to listen for
register messages and identifies the service that sent the
request. However, the invalid call is still also logged by
the widget.

This PR fixes this by checking for InvalidBusName exceptions
that are caused when trying to introspect the object path
rather than the service and, where this happens, removing
the invalid item from the watcher.

This has no impact on the operation of the widget but will
reduce error messages in the log.