Fix CI failures
CI is failing with 2 errors:
1) cairocffi is not found
2) mypy errors

It looks like this might be related to an update to setuptools.

While we could pin setuptools, it seems we have been using deprecated
methods (e.g. `python setup.py install`) so I prefer to fix the issues.

This PR fixes the CI failures by:
1) Adding cairocffi installation to tox.ini
2) Installing qtile via `pip install .` in tox mypy environment
3) Fixing 2 mypy errors (unreachable code in utils.py)