Reset colours in Chord widget
Chord widget sets colours based on entry in `chord_colors` dict.
If the chord is not in the dict then the colours are not changed.
This can lead to undesirable location if the widget needs to
occupy a fixed space in the bar but have a default colour when
not showing a chord name.

This PR resets the chord's background and foreground colours when
a chord is left or when a chord is not in the `chord_colors` dict.

Fixes #3461