Release `Drag` events when clicking on Bar
If a user clicks on the bar when pressing a mod key and has
a matching `Drag` event in their config then the `qtile._drag`
value is set with the bar's position.

While the `_drag` value is set, windows will not be added to a group
as this would prevent dragging windows. However, when clicking on the
bar, the button_release event is swallowed by the bar and never passed
back to the server so the `_drag` value is never unset.

This causes undesirable behaviour where new windows are not added to the
current group until the release event is triggered on the server.

This PR fixes the issue by making sure `core.process_button_release` is
called after then window's own `process_button_release`.

Fixes #3144