Battery widget show as full when `Not charging`
As per #3411, some batteries report their state as `Not charging`
when they are full. The widget doesn't recognise this and shows
the battery being in an "unknown" state.

While the issue can be solved by using the `unknown_char` parameter,
that's undesirable in case a different, unknown state were to arise.

This PR adds a config parameter `full_not_charging` which, when set,
will mark the battery as full when the state is `Not charging`.

Fixes #3411