Notify: fix release of dbus service in `finalize`
There is a bug where a user has the `Notify` widget installed and
tries to do `reload_config`. The result of the bug is that the
notification server is unavailable fter reload.

As the release of the service name during `Notify.finalize()` was
an async call, it may not complete until after the config has been
reloaded. However, because the server was in place when the config
was reloaded, a new server is not started. The `finalize` method
then removes the server and the user is left with no server.

This PR fixes the issue by ensuring that the release of the
service name is completed synchronously as part of `finalize`.

Fixes #3475