WL: restack xwayland clients to top of Z axis upon activation
This Z axis management is 'below' the Qtile level, and pertains to how
they are seen from the POV of XWayland. While we may not have XWayland
clients mapped at the same time, the XWayland considers them to be, so
an unmapped window can actually obscure a mapped one and soak up the
input.

This requires the next release of pywlroots.

Fixes #3280