WL: Don't use unknown variable when finding output for xdg_popup
The `Output` variable here is not imported at runtime, so a NameError is
always raised at this assertion. This wasn't caught because at testtime
it is loaded. Gah.

As of yet, backend exceptions are not caught and logged. This should be
fixed so that any exceptions raised in the wayland backend internal code
can be identified.