Log to either stdout or file
This sets the logger to either log to stdout (tests/xephyr) or a file,
but not both. During regular usage, printing to stdout is pointless.
During tests/dev, logging to file is noise (assuming one is within a
regular qtile session). This sets the default to stdout to simply its
usage (init_log is called much more in tests). This also removes
log_color and ties it to stdout so it does print in colour to stdout,
but not to log.