window: add an is_visible() method
motivation for this is wanting the traverse.py implementation in
qtile-examples to not traverse hidden windows. With this patch, the
following diff does not traverse windows that have been hidden by the
layout. (I'll send a patch to traverse.py when we land whatever the final
API is here.)

diff --git a/dotfiles/qtile/traverse.py b/dotfiles/qtile/traverse.py
index 5b693af..298bc1a 100644
--- a/dotfiles/qtile/traverse.py
+++ b/dotfiles/qtile/traverse.py
@@ -70,7 +70,7 @@ def _focus_window(qtile, dir, axis):
         windows.remove(cur)

     for w in windows:
-        if isinstance(w, Screen) or not w.minimized:
+        if isinstance(w, Screen) or (not w.minimized and w.is_visible()):
             pos = getattr(w, axis) + getattr(w, dim) / 2
             gap = dir * (pos - cur_pos)
             if gap > 5:

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>