WL: Always raise menu-like windows to top of xwayland Z stack
Currently XWayland windows are only raised to the top of the Z stack if
they are normal windows (non override-redirect) or appear to want focus.
wlroots uses an estimate to determine which OR windows want focus. Some
windows (e.g. Ardour menu windows) seem to fall through the cracks of
this estimate, appearing like normal windows but are labelled as OR. If
that happens we should still raise them to the top for expected
functionality.