Fire client_killed hook after client is removed from group
I'm not sure why this isn't the case already but let's keep an eye out
for issues that arise due to this. We want a killed window to be removed
from its group so that hooked functions that cause the group to lay
out and configure windows don't cause errors (e.g. a wlroots
assertion is raised when configuring a killed window). However we want
`c.group` to still point to the group because it's useful for hooked
functions, and code already exists that use it so this stops that from
breaking.

fixes #3513