Fire client_killed hook after client is removed from group
I'm not sure why this isn't the case already but let's keep an eye out
for issues that arise due to this. I suspect most uses of this hook
don't involve the group but those that do may see different
behaviour.

This fixes a Wayland bug where laying out the group's layout with a
killed window raises a wlroots assertion.

fixes #3513