Handle custom layout names in CurrentLayoutIcon
Users may specify custom names for layouts in their configs. This
can cause the `CurrentLayoutIcon` to fail as it's looking for
an icon called `layout-custom-name` which may not exist.

This PR addresses the issue by enabling the widget to look for
an icon for the underlying layout class name.

Fixes #3656