WL: add commands to hide and unhide the cursor
This functionality is needed for feature parity with X11, but
unfortunately is not (yet) available through a protocol, and so third
party utilities cannot do this for us.

This adds two commands:

- core.hide_cursor: hides the cursor until it is unhidden again
- core.unhide_cursor: unhides it, if it was hidden

Example usage would be via a tool such as swayidle to hide the cursor
when input has idled for a while, and then unhide the cursor when input
starts again.

Closes #3651