Fix libffi version issue in CI
New release of cffi (1.15.1) seems to break cairocffi. For pypy
we pin the version at 1.15.0 but for other python version we
can force a rebuild of cffi to use the installed libffi library
rather than the precompiled wheel.

We also need to pin the cffi version in the readthedocs build so
that widget images can be rendered.