Better support for multiple StatusNotifier widgets
The StatusNotifierHost only allowed a single callback to be registered.
This prevents different widgets from registering their own callbacks.
This is not particularly problematic for the stock widget but does
cause issues for the qtile-extras version which needs to draw context
menus as the menu could be rendered against the wrong instance.

This PR addresses the issue by appending registered callbacks to a list
and calling each whenever the triggering event arises.

Fixes #3738