fix: log python errors in config.py also on startup
Previously, any syntax errors in `config.py` were silently ignored
on startup due to the exception of the module import being caught
and converted into a simple notification message.

This makes it hard to debug erros in the config file. On reload,
this already is logged as expected, just not on the initial start.

This PR adds the exception log on startup to make the behavior
consistent in both cases.