WL: refactor inputs and add command to list input information
This refactors the way the wayland backend manages and configures input
devices. Mostly this involved moving around code in inputs.py into more
of an OO structure around the existing `Keyboard` and a new `Pointer`
class, and associating the device configuration code with those classes.
Then the `Core` also keeps track of `Pointer`s as it does already with
`Keyboard`s. IMO this makes the code easier to understand, but it also
makes it possible to get device information on demand. This is taken
advantage of with a new `Core.cmd_get_inputs` command, which will return
information on all existing input devices. This is equivalent to sway's
`swaymsg -t get_inputs` which is very useful for getting identifiers for
input devices that make it possible to target `InputConfig` options to
specific devices.