WL: update input configuration upon config reload
This converts `Core.distribute_windows` into a more generic
`Core.on_config_load` which is called when the config is loaded. The
core can then do what it needs to do with the updated configuration i.e.
organise windows (as before) or configure input devices (supported by
this change in wayland).