WL: refinement of pointer state changes
This makes the wayland core track which window of any type is under the
pointer, rather than just internal windows. This lets us detect when the
pointer moves out of a Window/Static window, so we can reset the cursor
image (to the default 'left_ptr' icon) when that happens. Previously the
icon was set upon all cursor movements if no non-internal window was
under the pointer, doing needless work.

This also prevents windows that don't have pointer focus from changing
the icon (firefox does this), and improves handling of focus and icon
when transitioning between windows and borders.