WL: remove some unnecessary width/height casting and getting
Some `window.place` calls are explicitly calling `int` on passed
geometry because in the past it was possible for floats to get passed
from the backend, but the method has since been typed with `int`s so the
cast is unnecessary. This commit also removes some unnecessary getting
of the geometry properties and gets the underlying private attributes.
This is borderline premature optimisation, but the properties only exist
to be compatible with the base classes and we're only accessing private
attributes, so, why not 😄