WL: position cursor and set default image after backend starts up
The wlroots backend (i.e. `Core.backend`) sets up any input devices that
are connected at startup but for some reason the cursor image doesn't
stick at this time. Setting the default cursor image after it's started
(and input devices have been set up) seems to fix this. The warping
places the cursor at (0, 0) as default. wlroots seems to place it at
(100, 100), but we draw the cursor at (0, 0) initially so it 'jumps'
after the very first movement. This fixes that too.