Border color for stacked windows in stack layout
This adds a new feature in stack layout. Now it is possible to change border color if the windows are in 'split' or 'stacked' state. This feature is present in column layout and is very useful to quickly see if there are windows "hiding" below. The feature will be as useful in stack layout.