WL: return shell in window info, add helpers for wayland-specific testing
This adds a `"shell"` item to the dictionary returned by the `info`
window command. This will be one of "XDG", "XWayland" or "layer". Mostly
useful for debugging.

This also adds a test for this behaviour. This needed first adding a
couple of test helpers for test/backend/wayland like with x11. These
will also be used for future wayland tests.