Simplify running lazy command from config
We often see users trying to run lazy commands from within a function in
their config. This is because there is to confusion about what lazy
commands actually do. In addition, while users are familiar with the
lazy command syntax, they are less familiar with the underlying python
objects.

This PR addresses these issues by adding a `qtile.run_lazy` method which
allows users to pass a lazy command as an argument and this will be run.