Add group sliding implementation to Wayland backend
Here the core starts a group slide by positioning windows in preparation
for the sequence, adjusts positions interactive during the sequence, and
then at the end of the interactivity (via
`Qtile.process_button_release`), it will smoothly transition to the
target group (i.e. if the drag/swipe moved over 50% of the screen's
width in either direction it'll be a new group). So this finishes
with a non-interactive animation to avoid jarring jumps.