Fix Notify `CloseNotification` bug
`CloseNotification` tried to look up the notifcation by the id received
by the method. However, the id was not zero indexed and so sending the
ID of the last notification resulted in an IndexError.

This lookup was also redundant as the callbacks require the notification
ID rather than the notification itself.

Fixes #3867