Add decorator to hide commands
By default, when an object inherits from a CommandObject, any methods
that were exposed on the command graph will also be exposed for the
child, regardless of whether the method is explicitly decorated or not.

Sometimes, we may want methods to be hidden (e.g. when they are not
appropriate for the child object). This PR adds a decorator that will
hide methods from the graph.