drop support for python 3.8
pyflakes 3.0 dropped support for python2.x-style comment type annotations,
so we get a lot of F401s, which the next patch will fix.

We can't use dict[] subscripting until 3.9: https://bugs.python.org/issue45117

So we could either switch to typing.Dict which is more verbose, pin
pyflakes, or drop support for 3.8. Let's move to the future and drop 3.8.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>