Warn when config duplicates Key mod+key spec
An alternative would be to ignore the duplicate (keep the first), or
support specifying multiple `Key`s and keep lists in `self.keys_map`
(which is already what `self.mouse_map` does, so one can actually
specify multiple `Mouse` instances which get sequentially triggered).

However, as is this does not change behaviour.