Fix simple_key_binder break after KeyChord
Currently, after using a KeyChord, users using the simple_key_binder are
unable to change groups. When grabbing a chord, all
other keys are ungrabbed. However, when the chord is exited, keys are
regrabbed but this is done by reference to the keys variable in config
which does not include the keys defined by the simple_key_binder.

This PR fixes the issue by appending the keys to the qtile.config.keys
variable.

Fixes #4024