more auto_minimize hackery
when steam games lose focus, they try to:

1. set IconicState
2. remove _NET_WM_FULLSCREEN

previously with the auto_minimize behavior we ignored the first part, which
was fine. If a user only had one monitor, when _NET_WM_FULLSCREEN is
removed, we toggle_floating() and re-place() the window according to where
the bar was. Except that people with one monitor didn't see this behavior.

If users have multiple monitors, this removal of _NET_WM_FULLSCREEN results
in a toggle_floating() which moves the fullscreen window to not being
fullscreen, which makes it pop into the current layout, which is weird.

Instead, let's not respect _NET_WM_FULLSCREEN removals like we're not
respecting setting to IconicState, and get rid of both of these problems.

Maybe this is a hack-on-hack, and we should find some other way to
implement this. I can't really tell myself, but I don't want things to
auto minimize just because I switch screens.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>