Relax regex on qtile shell command arg parsing
Qtile shell currently uses a regular expression to split the command
name from the arguments. However, the arguments only match a limited
subset of characters; characters like `.`, `[]` etc are not matched
resulting in an `InvalidCommand`. This limits the usefulness of the
`eval` command in particular.

This PR fixes this by allowing any characters to be passed as args.