Updates to Gap/Bar margin and strut handling
Changes:

- wayland layer surfaces correctly update their reserved space
- Gaps use margins to track reserved space, rather than mutating their
  size. When reserving space on a screen edge with no gap/bar, the Qtile
  manager will create a new Gap with size 0.
- Bars also make use of the Gap code for calculating geometry with
  margins, so Bar._configure is updated to update its geometry only with
  reserved space and border widths. Gap.margin is therefore mutated by
  Qtile to manage reserved space, whereas Bar.margin is *not* mutated
  and always reflects the config option passed to it.
- Bar.size and Gap.size is also made immutable by hiding it behind a
  setter-less property and _size.
- Bars check valid border colors in __init__
- added utils.is_valid_colors helper to check valid colors (probably
  useful elsewhere)
- Gap and Bar method type signatures are annotated
- A couple Bar attributes are dropped, for example crashed_widgets which
  only needs to exist during _configure
- Some attributes are renamed to either reflect their private usage, to
  be consistent (i.e. `_cursor_in` -> `_has_cursor` to match
  `_has_keyboard`), to be more clear (all mentions of 'strut' have
  changed to just 'reserved space' as the former is an X-specific word).
- Bar.queued_draws doesn't need to actually count the queue, just that
  there is one (changed to bool)

And lastly the ugly bit - in an attempt to see how to type
`Configurable` classes, annotations are added for some options (i.e.
their respective attributes). See Bar.__init__. There is a lot of room
for improvement for typing configurables but designing a solution that
helps us rather than hinders us is critical, and even though I've tried
a few approaches in the past I'm still not happy with the solution. The
approach here is the most basic one: annotate the attribute in
__init__ and tell mypy to ignore any expressions that seem to contradict
that type. You can see this here with `self.margin` in `Bar.__init__`.
This is made more complicated by having a slight different type for the
`margin` (`int | list[int]`) option than the `margin` attribute
`list[int]`. The class having this difference is fine - so any approach
to typing configurables has to handle this neatly.