Fix width issue for scrolling text
Currently, when widgets have `scroll=True` set, the `width` parameter
acts as maximum width. Where the width of the text is less than this
value, the widget will shrink accordingly. This may be undesirable for
users wishing to have a fixed width at all times.

This PR adds a new parameter, `scroll_fixed_width`, which, when set to
`True`, forces the text widget to have a fixed width at all times.

Fixes #4101