Fix `PangoLayout` error on widget reload
We experience `PangoLayout` errors if widgets try to draw to drawers
that have been finalised. This can happen when timers schedule drawing.
We deal with this by making sure timers are cancelled when widgets are
finalised. However, we only finalise timers that have been created by
`self.timeout_add`. Timers set by `qtile.call_soon/later` are not
tracked. These method are used to set initial timers for widgets at the
end of `_configure` so, if these timers have not yet triggered before
the widget is finalised then they will still be triggered.

This PR fixes that issue by ensuring that initial timers are also set
via `timeout_add` so that they can be cancelled on finalising the
widget.

Fixes #3869