Remove Heisenbugs
As mentioned by @tych0, there are potential "Heisenbugs" in our code
base where we use `asyncio.create_task` but don't retain a reference to
the created task. In those scenarios, the garbage collector could
collect the task before it's executed.

This PR stores created tasks in a list and periodically removed
completed tasks from that list to limit its size.