Fix Notify widget finalisation
The `Notify` widget was failing to remove callbacks when the widget was
finalised. This meant that callbacks were triggered on widgets that no
longer had an active `Drawer` resulting in error messages.

This PR simplifies the finalisation process by removing the call for the
notifier to release the service name from bus. The name will be released
automatically when qtile shuts down so there is no need to explicitly
call the method.

Fixes #4153