Don't mask AttributeErrors that arise during startup
When our argparsing runs the target subcommand, it does so inside a try
block which appears to have been added to catch the absence of
`options.func` when someone just runs `qtile`, rather than `qtile
start`. This has the side effect of hiding any AttributeErrors that
arise early on in startup e.g. when the core is being created. This
doesn't affect users so much (except those really living on the edge)
but it can be annoying during development. Instead, this does the
attribute check explicitly.

Fixes #4152