Fix StatusNotifier test
The `dbus_next` 0.2.3 release has a bug that causes a segmentation fault
when run in a Gtk mainloop on python 3.11. This bug is fixed in the git master but
there has not yet been a 0.2.4 release (and it's unclear when this
release will happen).

This PR includes a patch to fix the bug which enables us to run our
`StatusNotifier` tests. The patch should, however, be removed when
`dbus_next` publishes its next release.