Reworking migrations
This commit replaces the current `qtile migrate` implementation with a
new one. The previous version was based on the `bowler` library but this
library is not compatible with future versions of python and so may
break in the future.

The new version is based on `LibCST`.

As part of the reworking, I have added a "simple" framework for
adding new migrations, supporting documentation and  tests. Everything is contained in a
single file.