fixed double margin problem in the columns layout
In order to avoid the issue of duplicated margins, I had to remove the
option to set distinct margins for different sides of the window. I
don't really see what the benefit of exposing this to the user would be
anyway.

With this patch, the user will be able to set the margin to a number and
have that be the gap between the windows and between the outermost
windows and the edge of the screen. I think this is the behaviour the
majority of the users expect the margin argument to have.

I also updated and added some tests for the columns layout.