Config: Set group after configuring bar
When we do not do this then the layout.show() method gets a screen_rect
that does not account for the bar borders. So we first configure the gaps
(bars) such that the screen gets the correct screen_rect when used in
set_group that further calls group.set_screen, which calls layout.show()

You can observe this issue with the treetab layout. Start qtile with a bar
that has borders and have the treetab be the first layout to be shown.
You can see that the heading is not set correctly, when you spawn a window
or switch to and back to the group it fixes itself.