Add suspend hook
Adds a new `suspend` hook which is called before system goes to sleep.
This depends on systemd's logind service so users need both systemd and
dbus-next installed for this to work.

Where a user sets a suspend hook, qtile will create a sleep inhibitor
that delays (NB not blocks) suspend until the hook has completed.