Group + Manager: Set screen_affinity properly on startup
We had screen_affinity defined to "make sure the group starts on group
x" but this was not actually correctly used on startup. The initially
focused group did not take the affinity setting into account.

On startup we should also check the screen affinity and properly
create new groups if no available group can be found for the
screen. This should fix #3880