Fix ability to erase `Drawer` background
There is an apparent issue/limitation with the cairo library where
setting the `OPERATOR_CLEAR` library in a `RecordingSurface` will not
clear the background when painted to an `XCBSurface`. This means that,
for our x11 backend code, we need to clear the underying `XCBSurface`
directly.

To simplify this, our `Drawer.clear` method (which actually clears and
paints a background colour) has been updated to include a call to a new
method, `clear_rect`, which does the erasing of the background. For X11,
this new method clears the `XCBSurface` rather than the recording
surface.

The new method also accepts optional parameters to allow us to clear a
subset of the `Drawer`. This is currently only used by third_party
libraries (i.e. qtile-extras) and should help them avoid the need for
writing backend-specific code.