Fix timers on dead widgets
Some widgets may try to update their text layout after the widget has
been finalised. This will result in errors in the logfile as the pointer
in the PangoLayout has been set to None. We should therefore check that
the layout has not been finalized before trying to update it.

In addition, I have added code to prevent timers being set on finalised
widgets.