PulseVolume widget: move PulseAsync creation time (#4428)
The PulseAsync object is currently initialised when the widget is initialised.
As part of this step it looks for the current eventloop. That's fine when the
config is loaded by qtile but it will break if a user runs python config.py to
check their config for errors.

This moves the initialisation of the object to _configure.