Fix two bugs when adding/removing screens
There are two bugs that can arise when a screen is
disconnected and plugged back in:

1) Disconnecting the screen calls `kill_window()` on the gaps on that
   screen. However, that method does not include a call to cancel any
scheduled timers (e.g. `self.draw()`) so the bar my try to draw
finalised widgets resulting in a PangoError when the bar tries to get
the size of the widgets in `self._resize()`.

2) Reconnecting a screen could result in widgets on that screen not
   updating. This was because the screen uses previously
initialised/configured widgets and so `self.configured` and
`self.finalized` were both `True`. This prevented widgets from being
reconfigured correctly as well as blocking any timers.