Fix layout issue with killed scratchpads
When a scratchpad is killed (rather than toggled) it can prevent layouts
from setting layout correctly when tiled windows are closed.

I'm not entirely sure why this happens but the bug is fixed by having
`layout._ClientList.remove` return the next window which allows the
layout to focus the window correctly.