WL: print message at wayland core.py ffi module no import (#4467)
Adds a slightly better message when this file is imported from but `lib`
couldn't be loaded. We allow this so that the docs env doesn't need
pywlroots. But sometimes if the core is used a `NameError` is raised
when accessing `lib` at qtile startup, and it isn't clear why.

E.g. see #4450