[x11] Add pseudo-transparency
Adds a new config option, `x11_fake_transparency`, to emulate
transparency in bars/widgets by copying the root window contents to the
background of the bar/widget before rendering the contents on top.

Primary benefit of this is to prevent issues with the `Systray` widget
which cannot currently be displayed with semi-transparent backgrounds.