Prevent KeyError in third party hook registries
Reloading the config can cause third party hook registries to raise a
KeyError when firing hooks. This is because the `subscriptions` dict is
cleared but the regitry key is only created when a hook is subscribed
(as the hook module is not reloaded). Therefore, if a user does not
subcribe to any hooks from the third party library, this will cause the
hook to fail.

Fixes: https://github.com/elParaguayo/qtile-extras/issues/301