Config: Do not match on substrings
For some reason in the Match class we match on substrings. So e.g. for
wm_class if value = ["test1", "bogus"] and our candidate is "test1what" we
would loop through the list and determine that the test passes because
test1 is a substring of test1what. This leads to confusion and bugs
such as #4602 and #2670.

In the current situation even the default config is pretty broken on
X11 with the float rules (e.g. wm_type dialo would
match dialog and thus result in a floating window).