improve ratio calculation of ratioTile layout
This was the initial description for a pull request that
did not work. But here it is suitable to explain the reasoning behind it.
I added a RatioTile(ratio=1) object in my config and noticed that the
layout favoured 3 upright elongated windows on a screen that is 1.125 higher than wide.
So they are really really slim, 1.125 tall and 0.333 wide.
The ratio is 1.125/(1/3)=3.375
But I expected them to be 3 wide "landscape" windows because that would be
1.125/3 tall and 1 wide, which is a better ratio because if you see the ratio
as longer_edge_of_window/shorter_edge_of_window then it is
1/(1.125/3)= 2.666

This patch can fulfill the described wish,
while keeping overall expected "ratioTile" behaviour.
With this patch a GOLDEN_RATIO=1 tries to make the windows "square".
If a slightly more landscape tendency is preferred,
a default GOLDEN_RATIO of 1.2 probably would work.