Fix duplicate key message with dgroups_key_binder
The dgroups_key_binder creates key bindings for the groups and adds
these keys to the config `keys` variable (this was needed to ensure keys
were correctly grabbed when exiting a KeyChord - see 9d65ae2e8f0a5283506f7db0d38ee14544fb5067) but this results in multiple attempts to bind the same key.
This PR silences warnings when the same key definition is passed.

Fixes #4651