Delete non-persistent groups when windows are moved to another group
Fix #4654

Save a reference to the original group before removing it from the
client and set it back before calling DGroups._del(client) so it can grab
the reference and schedule it for deletion.

Check if the client has already been scheduled for deletion in order to
avoid scheduling the same client multiple times and generating a
KeyError.