Resolve dbus sender name when handling signals
0b46ce4b3a2a220d0c84737fe60949cee8c81f4a introduced a fix to only
forward signals if they match the criteria. This was helpful for widgets
tracking multiple signals as it removed the need to include logic in
those widgets.

However, an issue arises if the widget is filtering my the sender of the
signal (e.g. Mpris2 widget where the user is only listening to a
specified player). In that instance, the filter is asking for a string
of the sender's pretty name (e.g. "org.mpris.MediaPlayer2") but the
signal includes the sender's unique name (e.g. ":1.175") so there is no
match.

This PR fixes this issue by resolving the pretty name to the bus's
unique name.