[X11] Bring windows activated by pager to front
Qtile follows spec for stacking windows. However, this spec itself doesn't provide a way for window switcher to change the stacking order.
That's why most EWMH-compatible WMs in response to _NET_ACTIVE_WINDOW not only change focus, but also raise the window to the top of the stack.

Qtile already has code for handling activation by a pager so we can just
add code to bring the window to the front.

Closes #4670