Fix background bug for mirrored widgets
Since the refactoring of the `Drawer` class, widget backgrounds are now
drawn to the drawer's RecordingSurface. However, `Mirror.draw` paints
the widget background with the background colour before copying the
contents of the recording surface. This means that the background is
painted twice, causing issues where widgets had semi translucent
backgrounds.

This PR fixes the issue by clearing the drawer before copying the
recording surface.

Fixes #4589