Fix bugs in Mirror widget
The Mirror widget has two bugs:
1) Static width widgets do not have their width copied to the reflection
2) Negative width widgets break the widget.

The use case for point 2 is a bit of a hack: a `Spacer` with a negative
width will not be drawn as the widget is only drawn where the width is
greater than zero. However, the length is taken into account when
positioning other widgets in ths bar.

This PR fixes both issues by setting the width and preventing the widget
from drawing when it has a width <= 0.