Add padding to BatteryIcon
The existing padding was broken on `BatteryIcon`. It was 0 when the
images were loaded/resized and is then set (to a fixed value)
afterwards. However, this value was never used when caluclating the
widget's width or the icon's position.

This PR replaces the old method with a new, user-defined `padding`
parameter.